carton_yolo_project/requirements.txt
ultralytics>=8.3.0
PyYAML>=6.0.1
